+++ tests/testthat/test-plot_feature_similarity.R
@@ -65,6 +65,30 @@ familiar:::test_plots(
 )
 
 
+# Plots with selection of features.
+familiar:::test_plot_ordering(
+  plot_function = familiar:::plot_feature_similarity,
+  data_element = "feature_similarity",
+  outcome_type_available = c("continuous", "binomial", "multinomial", "survival"),
+  plot_args = list(
+    "features" = c("feature_1", "feature_2a", "feature_2b")
+    ),
+  debug = debug_flag
+)
+
+# Plots without tick labels on x and y-axes.
+familiar:::test_plot_ordering(
+  plot_function = familiar:::plot_feature_similarity,
+  data_element = "feature_similarity",
+  outcome_type_available = c("continuous", "binomial", "multinomial", "survival"),
+  plot_args = list(
+    "remove_feature_labels" = TRUE
+  ),
+  debug = debug_flag
+)
+
+
+
 # Test plotting from dataObject.
 data <- familiar:::test_create_good_data(outcome_type = "continuous")
 p <- familiar::plot_feature_similarity(object = data, feature_similarity_metric = "spearman")
@@ -106,5 +130,18 @@ p <- familiar::plot_feature_similarity(
   feature_similarity_metric = "spearman"
 )
 testthat::test_that("Plotting feature similarity using data.table works.", {
+  testthat::expect_true(is(p[[1L]], "gtable"))
+})
+
+
+
+# Test plotting from dataObject with set features.
+data <- familiar:::test_create_good_data(outcome_type = "continuous")
+p <- familiar::plot_feature_similarity(
+  object = data, 
+  features = c("feature_1", "feature_2a", "feature_2b"),
+  feature_similarity_metric = "spearman"
+)
+testthat::test_that("Plotting feature similarity using dataObject and a limited number of features works.", {
   testthat::expect_true(is(p[[1L]], "gtable"))
 })